+++ com_phpmyjoomla/administrator/views/managetables/tmpl/default_header.php
@@ -15,17 +15,30 @@ defined('_JEXEC') or die('Restricted access');
 $doc = JFactory::getDocument();
 
 $doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/datatables.min.css');
-$doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/dataTables.colVis.css');
+$doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/buttons.dataTables.min.css');
 $doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/dataTables.colReorder.css');
+$doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/keyTable.dataTables.min.css');
 $doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/jquery.modal.css');
 $doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/phpmyjoomla_css_custom.css');
 $doc->addStyleSheet(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/css/phpmyjoomla.css');
 ?>
 <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" type="text/css">
 <?php
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/jquery-3.3.1.js');
 $doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/datatables.min.js');
-$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/dataTables.colVis.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/dataTables.buttons.min.js');
+
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/export/buttons.flash.min.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/export/jszip.min.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/export/pdfmake.min.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/export/vfs_fonts.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/export/buttons.html5.min.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/export/buttons.print.min.js');
+
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/buttons.colVis.min.js');
 $doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/dataTables.colReorder.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/jquery.dataTables.columnFilter.js');
+$doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/dataTables.keyTable.min.js');
 $doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/jquery.modal.min.js');
 $doc->addScript(JURI::root() . 'administrator/components/com_phpmyjoomla/assets/js/phpmyjoomla_js_custom.js');
 
@@ -38,6 +51,6 @@ $doc->addScript(JURI::root() . 'media/system/js/core.js');
 <div id="ajax_shield" name="export_shield" class="clean_background"></div>
 <div id="ajax_loading" class="loading-invisible">
     <p>
-        <img src="<?php echo JURI::root() . 'administrator/components/com_phpmyjoomla/views/managetables/tmpl/assets/images/loading.gif';?>" alt="Loading" />
+        <img src="<?php echo JURI::root() . 'administrator/components/com_phpmyjoomla/assets/images/loading.gif';?>" alt="Loading" />
     </p>
 </div>