+++ com_phpmyjoomla/CHANGELOG.md
@@ -11,13 +11,28 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 ### Removed
 
+## [2.1.0] - 2020-04
+### Added
+- Excel and Print buttons
+
+### Fixed
+- Change Log modal scrolling.
+- Column visibility button fixed.
+
+### Changed/Update
+- Updated css files.
+- Updated js files.
+
+### Removed
+- Cleaned code.
+
 ## [2.0.1] - 2018-11
 ### Added
 - Added plugin quick icon.
 - Package plugin and component.
 
 ### Removed
-- Clean code
+- Cleaned code
 
 ## [2.0.0] - 2016-12
 ### Changed/Update
@@ -34,7 +49,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added compatibility icons for joomla (3.x, 3.5, 3.4, 3.3) and php (7.0, 5.6, 5.5, 5.4).
 
 ### Removed
-- Clean code
+- Cleaned code
 
 ## [1.5.1] - 2016-01-11
 ### Changed/Update